Refactor : OAuth strategies into factory
Move provider-specific Passport strategies into a generated factory to
remove duplication and enable dynamic strategy creation. Remove many
unused OAuth DTOs and the user.types, simplify OAuth-related interfaces
and loosen profile type guards to use any. Update imports and module
wiring.
