+++ server/src/modules/auth/auth.module.ts
@@ -3,12 +3,12 @@ import { PassportModule } from '@nestjs/passport';
 import { ConfigModule } from '@nestjs/config';
 import { AuthService } from './auth.service';
 import { AuthController } from './auth.controller';
+import { DatabaseModule } from 'common/database/database.module';
 import {
   DiscordOAuthStrategy,
   GoogleOAuthStrategy,
   GitHubOAuthStrategy,
-} from './strategies/oauth.strategy';
-import { DatabaseModule } from 'common/database/database.module';
+} from './strategies/oauth-strategy.factory';
 
 @Module({
   imports: [
