+++ server/src/modules/auth/interfaces/oauth.types.ts
@@ -22,57 +22,10 @@ export interface NormalizedOAuthProfile {
   raw: any;
 }
 
-export interface DiscordRawProfile {
-  id: string;
-  username: string;
-  discriminator: string;
-  avatar: string | null;
-  email?: string;
-  verified?: boolean;
-  locale?: string;
-  mfa_enabled?: boolean;
-  flags?: number;
-  premium_type?: number;
-}
-
-export interface GoogleRawProfile {
-  id: string;
-  displayName: string;
-  name?: {
-    givenName?: string;
-    familyName?: string;
-  };
-  emails?: Array<{ value: string; verified?: boolean }>;
-  photos?: Array<{ value: string }>;
-  _json?: any;
-}
-
-export interface GitHubRawProfile {
-  id: string;
-  username?: string;
-  login?: string;
-  displayName?: string;
-  name?: string;
-  emails?: Array<{ value: string; verified?: boolean }>;
-  photos?: Array<{ value: string }>;
-  avatar_url?: string;
-  _json?: any;
-}
-
-export type RawOAuthProfile =
-  | DiscordRawProfile
-  | GoogleRawProfile
-  | GitHubRawProfile;
-
 export interface OAuthTokens {
   accessToken: string;
   refreshToken: string;
   expiresIn?: number;
-}
-
-export interface OAuthData {
-  profile: NormalizedOAuthProfile;
-  tokens: OAuthTokens;
 }
 
 export interface AuthenticatedUser {
@@ -88,19 +41,7 @@ export interface AuthenticatedUser {
 
 export type OAuthValidationResult = AuthenticatedUser | null;
 
-export interface ProviderStrategyConfig {
-  name: OAuthProvider;
-  strategyClass: any;
-  config: OAuthProviderConfig;
-}
-
-export interface ProfileNormalizationOptions {
-  provider: OAuthProvider;
-  requireEmail?: boolean;
-  defaultAvatar?: string;
-}
-
-export function isDiscordProfile(profile: any): profile is DiscordRawProfile {
+export function isDiscordProfile(profile: any): profile is any {
   return (
     typeof profile === 'object' &&
     'username' in profile &&
@@ -108,7 +49,7 @@ export function isDiscordProfile(profile: any): profile is DiscordRawProfile {
   );
 }
 
-export function isGoogleProfile(profile: any): profile is GoogleRawProfile {
+export function isGoogleProfile(profile: any): profile is any {
   return (
     typeof profile === 'object' &&
     profile !== null &&
@@ -117,7 +58,7 @@ export function isGoogleProfile(profile: any): profile is GoogleRawProfile {
   );
 }
 
-export function isGitHubProfile(profile: any): profile is GitHubRawProfile {
+export function isGitHubProfile(profile: any): profile is any {
   return (
     typeof profile === 'object' &&
     ('login' in profile || 'username' in profile) &&
